rubocop fix and new modules
