+++ src/app.rb
@@ -1,7 +1,8 @@
 require_relative './Classes/movie'
 require_relative './Classes/author'
 require_relative './Classes/db'
-require_relative './modules/app_module'
+require_relative './modules/create_items'
+require_relative './modules/show_items'
 
 class App
   include CreateItems
